requirements.txt
anyio==4.0.0
beautifulsoup4==4.10.0
certifi==2023.7.22
charset-normalizer==2.0.12
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
python-telegram-bot==20.0
PyYAML==6.0.1
requests==2.26.0
rfc3986==1.5.0
schedule==1.1.0
sniffio==1.3.0
soupsieve==2.5
urllib3==1.26.18
